Refactoring: Add data collection methods that can be specialized
We still collect the table data in the same shape as before, but
encapsulated that data collection in triplets of methods such as
"init_row()/process_field()/finish_row" and
"init_table()/process_row()/finish_table().  The idea is that these
methods can be overridden in specific subclasses, for example to
organize/process the data in a specific way.